Database: Update some tables after client's revision
Add isFarmer column to employees table to identify the employees that
using front-office is a farmer.

Then add longitude, latitude and price to save source of location based
on coordinate and price of transaction.

Change column name in product transactions from 'received_weight' to
'delivered_weight' and change enum value of payment_status from
'not delivered' to 'undelivered'.

Add customer_id column to product_transaction_images table to identify
owner of the image.

Split price column in products table to 'buy_price' and 'sell_price'
because buy and sell price is different.

Add index and not null to some columns.

Signed-off-by: Taufiq Pohan <m.taufiq30s@hotmail.com>